projeto_arquitetura_computadores/etapa2/requirements.txt
cycler==0.10.0
kiwisolver==1.3.1
matplotlib==3.4.2
numpy==1.21.1
Pillow==8.3.1
psutil==5.8.0
py-cpuinfo==8.0.0
pygame==2.0.1
pyparsing==2.4.7
python-dateutil==2.8.2
six==1.16.0
